requirements.txt
arrow==1.2.3
asgiref==3.5.2
binaryornot==0.4.4
black==22.10.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drfdocs==0.0.11
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
swagger-spec-validator==2.7.6
python-decouple==3.6
django-filter==22.1
faker
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
